[HER-17]Follow-up Move content component
